Add map layer legend to stories side panel
